requirements.txt
scikit-learn
scipy
numpy
tensorflow-gpu
pymysql